Moved types.json to a shared folder
