+++ src/.env
@@ -7,3 +7,4 @@ APP_EXPLORER_URL='https://explorer.dev.grid.tf/'
 APP_GRAPHQL_URL='https://graphql.dev.grid.tf/graphql/'
 APP_GRIDPROXY_URL='https://gridproxy.dev.grid.tf/'
 APP_NETWORK='dev'
+APP_VERSION='dev1.0.0'
